Move controllers to own package under org.oskari.spring
