+++ servlet-map/src/main/java/org/oskari/spring/controllers/ActionRouteController.java
@@ -1,4 +1,4 @@
-package fi.nls.oskari;
+package org.oskari.spring.controllers;
 
 import fi.nls.oskari.control.*;
 import fi.nls.oskari.log.LogFactory;
