+++ servlet-map/src/main/java/org/oskari/spring/controllers/MapController.java
@@ -1,4 +1,4 @@
-package fi.nls.oskari;
+package org.oskari.spring.controllers;
 
 import fi.nls.oskari.control.ActionParameters;
 import fi.nls.oskari.control.view.GetAppSetupHandler;
